refactor(agent): ACL 客户端收拢进 apps/agent/src/acl/，消灭单文件顶层目录
agent 对每个独立对端进程（kagami-llm / browser / spire / oss）的 HTTP 客户端此前散成
几个顶层单文件目录（spire/、browser/、oss/）。收拢进一个 apps/agent/src/acl/：

- spire-client / browser-client / oss-client / http-llm-client → src/acl/
- detect-mime（跨切面字节嗅探工具，非 ACL 客户端）→ src/common/
- 删空 spire/ browser/ oss/ 三个目录；llm/ 保留（playground service + handler）

纯移动 + import 路径段替换，零逻辑改动。client 文件自身相对 import 深度不变（均为 src/
直接子目录），一字未改；仅约 50 处消费方 import 的 `spire/spire-client → acl/spire-client`
之类段名替换。测试文件镜像迁移（browser-tools-golden 是工具测试、留原地）。

方案取舍：不抽 -client 包（各客户端只有 agent 一个消费者，为单消费者造 4 个包是仪式感），
收进 agent 内 acl/ 目录最省且直达诉求。metric 留在 @kagami/metric-client 包不动。

文档：ARCHITECTURE.md + README(zh/en) 的 agent 模块列表更新（加 acl/ + common/，删已迁包的
config/db/logger/auth 陈旧条目 + 迁走的 oss/，修 llm/ 描述）。

字节基线保绿：browser-client-wire（KV 缓存字节契约 #173）、oss-client 五错误码、spire-client。
build / typecheck / lint / format 全绿，test 1485 全过。

Refs #342
