+++ apps/agent/src/acl/spire-client.ts
