+++ apps/agent/src/agent/apps/amap/amap.app.ts
@@ -3,7 +3,7 @@ import type { App, AppStartupContext } from "@kagami/agent-runtime";
 import { renderServerStaticTemplate } from "@kagami/kernel/runtime/read-static-text";
 import { AmapClient } from "./client/amap-client.js";
 import type { RootAgentEffect } from "../../runtime/effect/root-agent-effect.js";
-import type { OssClient } from "../../../oss/oss-client.js";
+import type { OssClient } from "../../../acl/oss-client.js";
 import { GeocodeTool } from "./tools/geocode.tool.js";
 import { RegeocodeTool } from "./tools/regeocode.tool.js";
 import { SearchPoiTool } from "./tools/search-poi.tool.js";
