+++ apps/agent/src/agent/apps/amap/tools/static-map.tool.ts
@@ -4,7 +4,7 @@ import { AppLogger } from "@kagami/kernel/logger/logger";
 import type { AmapClient, StaticMapMarker, StaticMapPath } from "../client/amap-client.js";
 import { escapeAttr } from "../amap-screen.js";
 import type { RootAgentEffect } from "../../../runtime/effect/root-agent-effect.js";
-import type { OssClient } from "../../../../oss/oss-client.js";
+import type { OssClient } from "../../../../acl/oss-client.js";
 
 export const STATIC_MAP_TOOL_NAME = "static_map";
