+++ apps/agent/src/agent/capabilities/browser/tools/click.tool.ts
@@ -1,7 +1,7 @@
 import { z } from "zod";
 import type { ToolKind } from "@kagami/agent-runtime";
 import { BrowserToolComponent } from "./browser-tool-component.js";
-import type { BrowserClient } from "../../../../browser/browser-client.js";
+import type { BrowserClient } from "../../../../acl/browser-client.js";
 
 export const BROWSER_CLICK_TOOL_NAME = "browser_click";
