+++ apps/agent/src/agent/capabilities/browser/tools/observe.tool.ts
@@ -1,7 +1,7 @@
 import { z } from "zod";
 import type { ToolKind } from "@kagami/agent-runtime";
 import { BrowserToolComponent } from "./browser-tool-component.js";
-import type { BrowserClient } from "../../../../browser/browser-client.js";
+import type { BrowserClient } from "../../../../acl/browser-client.js";
 
 export const BROWSER_OBSERVE_TOOL_NAME = "browser_observe";
