+++ apps/agent/src/agent/capabilities/resource/application/resource-file.service.ts
@@ -3,8 +3,8 @@ import { mkdir, readFile, realpath, rename, stat, writeFile } from "node:fs/prom
 import os from "node:os";
 import path from "node:path";
 import { BizError } from "@kagami/kernel/errors/biz-error";
-import { detectMime } from "../../../../oss/detect-mime.js";
-import type { OssClient } from "../../../../oss/oss-client.js";
+import { detectMime } from "../../../../common/detect-mime.js";
+import type { OssClient } from "../../../../acl/oss-client.js";
 
 /** download_resource 的结果：落地的绝对路径 + 写入字节数。 */
 export type DownloadToFileResult = { absolutePath: string; size: number };
