+++ apps/agent/src/agent/capabilities/resource/application/resource.service.ts
@@ -1,5 +1,5 @@
 import { BizError } from "@kagami/kernel/errors/biz-error";
-import type { OssClient } from "../../../../oss/oss-client.js";
+import type { OssClient } from "../../../../acl/oss-client.js";
 
 /** 一份按 resId 解出的资源：原图字节 + MIME + 字节数 + 是否图片。 */
 export type ResolvedResource = {
