+++ apps/agent/src/agent/capabilities/spire/tools/choose.tool.ts
@@ -2,7 +2,7 @@ import { z } from "zod";
 import type { ToolKind } from "@kagami/agent-runtime";
 import { SpireToolComponent } from "./spire-tool-component.js";
 import { renderSpireScreen } from "../render/spire-screen.js";
-import type { SpireClient } from "../../../../spire/spire-client.js";
+import type { SpireClient } from "../../../../acl/spire-client.js";
 
 export const SPIRE_CHOOSE_TOOL_NAME = "choose";
