+++ apps/agent/src/common/detect-mime.ts
