+++ apps/agent/src/napcat/application/napcat-gateway/image-message-analyzer.ts
@@ -1,8 +1,8 @@
 import { AppLogger } from "@kagami/kernel/logger/logger";
 import { truncateWithEllipsis } from "@kagami/kernel/utils/text";
 import type { NapcatReceiveImageSegment } from "../../domain/napcat-segment.js";
-import { detectMime } from "../../../oss/detect-mime.js";
-import type { OssClient } from "../../../oss/oss-client.js";
+import { detectMime } from "../../../common/detect-mime.js";
+import type { OssClient } from "../../../acl/oss-client.js";
 import type { ImageAssetDao } from "../../infra/image-asset.dao.js";
 
 const logger = new AppLogger({ source: "service.napcat-gateway" });
